+++ Project-Omni-Ride-Network/Controllers/HomeController.cs
@@ -1,7 +1,12 @@
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Configuration;
+using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
+using System.Net;
+using System.Net.Mail;
+using System.Text;
 using System.Threading.Tasks;
 
 namespace Project_Omni_Ride_Network {
@@ -9,10 +14,15 @@ namespace Project_Omni_Ride_Network {
     public class HomeController : Controller {
 
         private readonly DataStore dbStore;
+        private readonly IConfiguration configuration;
 
         public HomeController(DataStore dbStore) {
             this.dbStore = dbStore;
             dbStore.EnsureDataStore();
+        }
+
+        public HomeController(IConfiguration conf) {
+            configuration = conf;
         }
 
         public IActionResult Index() {
@@ -96,10 +106,6 @@ namespace Project_Omni_Ride_Network {
 
         #region Page Information Routes
 
-        [Route(Routes.CONTACT)]
-        public IActionResult Contact() {
-            return View();
-        }
 
         [Route(Routes.KARRIERE)]
         public IActionResult Karriere() {
@@ -127,5 +133,66 @@ namespace Project_Omni_Ride_Network {
         }
 
         #endregion
+
+        #region Contact
+
+
+        [Route(Routes.CONTACT)]
+        public IActionResult Contact() {
+            return View();
+        }
+
+        [Route(Routes.CONTACT)]
+        [HttpPost]
+        public IActionResult Contact(ContactModel contact) {
+            if (ModelState.IsValid) {
+                var ourMail = "service@c-u-management.de";
+                var senderMail = contact.SenderEmail.ToString();
+                var subject = contact.Subject;
+                var mailText = new StringBuilder();
+                mailText.Append("Name: " + contact.SenderName + "\n");
+                mailText.Append("eMail: " + contact.SenderEmail + "\n");
+                mailText.Append(contact.Message);
+
+                try {
+                    MailerAsync(ourMail, senderMail, subject, mailText.ToString());
+                } catch (Exception ex) {
+                    return View();
+                }
+            }
+            return View();
+
+        }
+
+        #region HelperMethods
+
+
+
+        public async Task MailerAsync(string ourMail, string senderMail, string subject, string message) {
+            try {
+                using (var mail = new MailMessage()) {
+
+                    mail.From = new MailAddress(senderMail);
+                    mail.Subject = subject;
+                    mail.To.Add(new MailAddress(ourMail));
+                    mail.Body = message;
+                    mail.IsBodyHtml = true;
+
+                    using (var smtpClient = new SmtpClient(configuration.GetValue<string>("MailCredentials:Hostname"), configuration.GetValue<int>("MailCredentials:Port"))) {
+                        smtpClient.EnableSsl = true;
+                        smtpClient.UseDefaultCredentials = false;
+                        smtpClient.Credentials = new NetworkCredential(configuration.GetValue<string>("MailCredentials:Email"), configuration.GetValue<string>("MailCredentials:Passwort"));
+                        await smtpClient.SendMailAsync(mail);
+                    }
+                }
+            } catch (Exception ex) {
+                throw ex;
+            }
+        }
+
+        #endregion
+
+        #endregion
+
     }
 }
